broadmobi: add MM_PLUGIN_REQUIRED_QCDM flag
This fixes an issue where, with an BM818 connected over USB,
suspending/resuming the modem causes the modem to reset immediately.

The culprit was bisected to d6203265c2, where QCDM probing was disabled
unless explicitly requested, so I assume this modem needs it based on
the constant resetting when power management was enabled on the modem's
usb connection.