iface-modem-3gpp: don't compare profile-name on attach APNs
When a request to set a new eps bearer settings is received, ignore the
profile-name when comparing the previous configuration with the new one,
since the modem's profile might already have a name that won't match the
settings provided by the user. Profile names are used in QMI modems.
If the profile name does not match the existing one, the modem will
detach and reattach to the network.