broadband-modem-mbim: ignore registration set error if NwError=0 reported
Returning an error in the registration set request completely breaks
the async operation requesting the explicit registration, even if the
modem may have already been registered via indications while waiting
for the set request to be completed.

We now ignore the registration set error if it returns a generic
failure with NwError=0, as in certain Qualcomm based devices. We
ignore unconditionally, without explicitly checking if we're
registered or not, because the upper logic will anyway do that.

Fixes https://gitlab.freedesktop.org/mobile-broadband/ModemManager/-/issues/723