api,errors: add generic MM_CORE_ERROR_THROTTLED error
Will be used to report operations that are not run as they have been
attempted too many times already.

Modems will usually report this kind of error e.g. when attempting to
connect with the same invalid APN too many times. The modem at some
point decides that it won't attempt the APN connection with the
network any more, so it returns this error to the caller right away.