broadband-modem-qmi: don't complete the power update before receiving the response
When relying on the power update indications, we were completing the
operation as soon as the indication was received, regardless of
whether the request response was received or not.

Avoid doing that, and instead complete the operation only when both
indication and the response are received. This avoid misalignment in
the operation completion if a new one is scheduled just after the one
we have completed.