broadband-modem-qmi: change operator description priority
If an IoT SIM card is used, the service name is always used as the operator
description. This is not what is expected. Therefore the priority is
changed with this commit, so that the long name is tried to be set first,
then the short name and finally the service name.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>