mm-broadband-modem-qmi: always reload current operator description
When using a modem with the qmi plugin, the operator description is only
set once for the current modem object. This is the operator description
of the mobile network to which the modem has first established a connection.

The operator description will no longer be updated even though on a
reconnection with a different mobile network.

Normally this is not a problem, as the SIM card usually dials into the
home network and stays there. But if an IoT SIM card is used, then this
is a problem because the operator description is no longer updated.

To fix this, the 'current_operator_description' is always loaded on
'modem_3gpp_load_operator_name' regardless of whether an operator
description has ever been set.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>